+++ src/Cache.php
@@ -43,6 +43,8 @@
 
 namespace SimplePie;
 
+use SimplePie\Cache\Base;
+
 /**
  * Used to create cache objects
  *
@@ -81,8 +83,8 @@ class Cache
      *
      * @param string $location URL location (scheme is used to determine handler)
      * @param string $filename Unique identifier for cache object
-     * @param string $extension 'spi' or 'spc'
-     * @return \SimplePie\Cache\Base Type of object depends on scheme of `$location`
+     * @param Base::TYPE_FEED|Base::TYPE_IMAGE $extension 'spi' or 'spc'
+     * @return Base Type of object depends on scheme of `$location`
      */
     public static function get_handler($location, $filename, $extension)
     {
@@ -103,7 +105,7 @@ class Cache
      */
     public function create($location, $filename, $extension)
     {
-        trigger_error('Cache::create() has been replaced with Cache::get_handler(). Switch to the registry system to use this.', E_USER_DEPRECATED);
+        trigger_error('Cache::create() has been replaced with Cache::get_handler(). Switch to the registry system to use this.', \E_USER_DEPRECATED);
         return self::get_handler($location, $filename, $extension);
     }
 
@@ -111,7 +113,7 @@ class Cache
      * Register a handler
      *
      * @param string $type DSN type to register for
-     * @param string $class Name of handler class. Must implement \SimplePie\Cache\Base
+     * @param class-string<Base> $class Name of handler class. Must implement Base
      */
     public static function register($type, $class)
     {