+++ src/Cache/Base.php
@@ -59,21 +59,21 @@ interface Base
      *
      * @var string
      */
-    const TYPE_FEED = 'spc';
+    public const TYPE_FEED = 'spc';
 
     /**
      * Image cache type
      *
      * @var string
      */
-    const TYPE_IMAGE = 'spi';
+    public const TYPE_IMAGE = 'spi';
 
     /**
      * Create a new cache object
      *
      * @param string $location Location string (from SimplePie::$cache_location)
      * @param string $name Unique ID for the cache
-     * @param string $type Either TYPE_FEED for SimplePie data, or TYPE_IMAGE for image data
+     * @param Base::TYPE_FEED|Base::TYPE_IMAGE $type Either TYPE_FEED for SimplePie data, or TYPE_IMAGE for image data
      */
     public function __construct($location, $name, $type);