+++ src/IRI.php
@@ -152,7 +152,7 @@ class IRI
         if (method_exists($this, 'set_' . $name)) {
             call_user_func([$this, 'set_' . $name], $value);
         } elseif (
-               $name === 'iauthority'
+            $name === 'iauthority'
             || $name === 'iuserinfo'
             || $name === 'ihost'
             || $name === 'ipath'
@@ -489,13 +489,13 @@ class IRI
                 || $character >= 0xFDD0 && $character <= 0xFDEF
                 || (
                     // Everything else not in ucschar
-                       $character > 0xD7FF && $character < 0xF900
+                    $character > 0xD7FF && $character < 0xF900
                     || $character < 0xA0
                     || $character > 0xEFFFD
                 )
                 && (
                     // Everything not in iprivate, if it applies
-                       !$iprivate
+                    !$iprivate
                     || $character < 0xE000
                     || $character > 0x10FFFD
                 )
@@ -717,14 +717,17 @@ class IRI
         if ($iri === null) {
             return true;
         } elseif (isset($cache[$iri])) {
-            list($this->scheme,
-                 $this->iuserinfo,
-                 $this->ihost,
-                 $this->port,
-                 $this->ipath,
-                 $this->iquery,
-                 $this->ifragment,
-                 $return) = $cache[$iri];
+            [
+                $this->scheme,
+                $this->iuserinfo,
+                $this->ihost,
+                $this->port,
+                $this->ipath,
+                $this->iquery,
+                $this->ifragment,
+                $return
+            ] = $cache[$iri];
+
             return $return;
         }
 
@@ -739,14 +742,17 @@ class IRI
             && $this->set_query($parsed['query'])
             && $this->set_fragment($parsed['fragment']);
 
-        $cache[$iri] = [$this->scheme,
-                             $this->iuserinfo,
-                             $this->ihost,
-                             $this->port,
-                             $this->ipath,
-                             $this->iquery,
-                             $this->ifragment,
-                             $return];
+        $cache[$iri] = [
+            $this->scheme,
+            $this->iuserinfo,
+            $this->ihost,
+            $this->port,
+            $this->ipath,
+            $this->iquery,
+            $this->ifragment,
+            $return
+        ];
+
         return $return;
     }
 
@@ -794,10 +800,12 @@ class IRI
             $this->port = null;
             return true;
         } elseif (isset($cache[$authority])) {
-            list($this->iuserinfo,
-                 $this->ihost,
-                 $this->port,
-                 $return) = $cache[$authority];
+            [
+                $this->iuserinfo,
+                $this->ihost,
+                $this->port,
+                $return
+            ] = $cache[$authority];
 
             return $return;
         }
@@ -822,10 +830,12 @@ class IRI
                   $this->set_host($remaining) &&
                   $this->set_port($port);
 
-        $cache[$authority] = [$this->iuserinfo,
-                                   $this->ihost,
-                                   $this->port,
-                                   $return];
+        $cache[$authority] = [
+            $this->iuserinfo,
+            $this->ihost,
+            $this->port,
+            $return
+        ];
 
         return $return;
     }