+++ src/SimplePie.php
@@ -43,11 +43,14 @@
 
 namespace SimplePie;
 
+use SimplePie\Cache\Base;
+use SimplePie\Cache\Psr16;
 use SimplePie\Content\Detector;
 use SimplePie\Exception\HttpException;
 use SimplePie\HTTP\FileClient;
 use SimplePie\HTTP\FileResponse;
 use SimplePie\HTTP\Response;
+use Psr\SimpleCache\CacheInterface;
 
 /**
  * SimplePie
@@ -60,350 +63,350 @@ class SimplePie
     /**
      * SimplePie Name
      */
-    const NAME = 'SimplePie';
+    public const NAME = 'SimplePie';
 
     /**
      * SimplePie Version
      */
-    const VERSION = '1.7.0';
+    public const VERSION = '1.7.0';
 
     /**
      * SimplePie Website URL
      */
-    const URL = 'http://simplepie.org';
+    public const URL = 'http://simplepie.org';
 
     /**
      * SimplePie Linkback
      */
-    const LINKBACK = '<a href="' . self::URL . '" title="' . self::NAME . ' ' . self::VERSION . '">' . self::NAME . '</a>';
+    public const LINKBACK = '<a href="' . self::URL . '" title="' . self::NAME . ' ' . self::VERSION . '">' . self::NAME . '</a>';
 
     /**
      * No Autodiscovery
      * @see SimplePie::set_autodiscovery_level()
      */
-    const LOCATOR_NONE = 0;
+    public const LOCATOR_NONE = 0;
 
     /**
      * Feed Link Element Autodiscovery
      * @see SimplePie::set_autodiscovery_level()
      */
-    const LOCATOR_AUTODISCOVERY = 1;
+    public const LOCATOR_AUTODISCOVERY = 1;
 
     /**
      * Local Feed Extension Autodiscovery
      * @see SimplePie::set_autodiscovery_level()
      */
-    const LOCATOR_LOCAL_EXTENSION = 2;
+    public const LOCATOR_LOCAL_EXTENSION = 2;
 
     /**
      * Local Feed Body Autodiscovery
      * @see SimplePie::set_autodiscovery_level()
      */
-    const LOCATOR_LOCAL_BODY = 4;
+    public const LOCATOR_LOCAL_BODY = 4;
 
     /**
      * Remote Feed Extension Autodiscovery
      * @see SimplePie::set_autodiscovery_level()
      */
-    const LOCATOR_REMOTE_EXTENSION = 8;
+    public const LOCATOR_REMOTE_EXTENSION = 8;
 
     /**
      * Remote Feed Body Autodiscovery
      * @see SimplePie::set_autodiscovery_level()
      */
-    const LOCATOR_REMOTE_BODY = 16;
+    public const LOCATOR_REMOTE_BODY = 16;
 
     /**
      * All Feed Autodiscovery
      * @see SimplePie::set_autodiscovery_level()
      */
-    const LOCATOR_ALL = 31;
+    public const LOCATOR_ALL = 31;
 
     /**
      * No known feed type
      */
-    const TYPE_NONE = 0;
+    public const TYPE_NONE = 0;
 
     /**
      * RSS 0.90
      */
-    const TYPE_RSS_090 = 1;
+    public const TYPE_RSS_090 = 1;
 
     /**
      * RSS 0.91 (Netscape)
      */
-    const TYPE_RSS_091_NETSCAPE = 2;
+    public const TYPE_RSS_091_NETSCAPE = 2;
 
     /**
      * RSS 0.91 (Userland)
      */
-    const TYPE_RSS_091_USERLAND = 4;
+    public const TYPE_RSS_091_USERLAND = 4;
 
     /**
      * RSS 0.91 (both Netscape and Userland)
      */
-    const TYPE_RSS_091 = 6;
+    public const TYPE_RSS_091 = 6;
 
     /**
      * RSS 0.92
      */
-    const TYPE_RSS_092 = 8;
+    public const TYPE_RSS_092 = 8;
 
     /**
      * RSS 0.93
      */
-    const TYPE_RSS_093 = 16;
+    public const TYPE_RSS_093 = 16;
 
     /**
      * RSS 0.94
      */
-    const TYPE_RSS_094 = 32;
+    public const TYPE_RSS_094 = 32;
 
     /**
      * RSS 1.0
      */
-    const TYPE_RSS_10 = 64;
+    public const TYPE_RSS_10 = 64;
 
     /**
      * RSS 2.0
      */
-    const TYPE_RSS_20 = 128;
+    public const TYPE_RSS_20 = 128;
 
     /**
      * RDF-based RSS
      */
-    const TYPE_RSS_RDF = 65;
+    public const TYPE_RSS_RDF = 65;
 
     /**
      * Non-RDF-based RSS (truly intended as syndication format)
      */
-    const TYPE_RSS_SYNDICATION = 190;
+    public const TYPE_RSS_SYNDICATION = 190;
 
     /**
      * All RSS
      */
-    const TYPE_RSS_ALL = 255;
+    public const TYPE_RSS_ALL = 255;
 
     /**
      * Atom 0.3
      */
-    const TYPE_ATOM_03 = 256;
+    public const TYPE_ATOM_03 = 256;
 
     /**
      * Atom 1.0
      */
-    const TYPE_ATOM_10 = 512;
+    public const TYPE_ATOM_10 = 512;
 
     /**
      * All Atom
      */
-    const TYPE_ATOM_ALL = 768;
+    public const TYPE_ATOM_ALL = 768;
 
     /**
      * All feed types
      */
-    const TYPE_ALL = 1023;
+    public const TYPE_ALL = 1023;
 
     /**
      * No construct
      */
-    const CONSTRUCT_NONE = 0;
+    public const CONSTRUCT_NONE = 0;
 
     /**
      * Text construct
      */
-    const CONSTRUCT_TEXT = 1;
+    public const CONSTRUCT_TEXT = 1;
 
     /**
      * HTML construct
      */
-    const CONSTRUCT_HTML = 2;
+    public const CONSTRUCT_HTML = 2;
 
     /**
      * XHTML construct
      */
-    const CONSTRUCT_XHTML = 4;
+    public const CONSTRUCT_XHTML = 4;
 
     /**
      * base64-encoded construct
      */
-    const CONSTRUCT_BASE64 = 8;
+    public const CONSTRUCT_BASE64 = 8;
 
     /**
      * IRI construct
      */
-    const CONSTRUCT_IRI = 16;
+    public const CONSTRUCT_IRI = 16;
 
     /**
      * A construct that might be HTML
      */
-    const CONSTRUCT_MAYBE_HTML = 32;
+    public const CONSTRUCT_MAYBE_HTML = 32;
 
     /**
      * All constructs
      */
-    const CONSTRUCT_ALL = 63;
+    public const CONSTRUCT_ALL = 63;
 
     /**
      * Don't change case
      */
-    const SAME_CASE = 1;
+    public const SAME_CASE = 1;
 
     /**
      * Change to lowercase
      */
-    const LOWERCASE = 2;
+    public const LOWERCASE = 2;
 
     /**
      * Change to uppercase
      */
-    const UPPERCASE = 4;
+    public const UPPERCASE = 4;
 
     /**
      * PCRE for HTML attributes
      */
-    const PCRE_HTML_ATTRIBUTE = '((?:[\x09\x0A\x0B\x0C\x0D\x20]+[^\x09\x0A\x0B\x0C\x0D\x20\x2F\x3E][^\x09\x0A\x0B\x0C\x0D\x20\x2F\x3D\x3E]*(?:[\x09\x0A\x0B\x0C\x0D\x20]*=[\x09\x0A\x0B\x0C\x0D\x20]*(?:"(?:[^"]*)"|\'(?:[^\']*)\'|(?:[^\x09\x0A\x0B\x0C\x0D\x20\x22\x27\x3E][^\x09\x0A\x0B\x0C\x0D\x20\x3E]*)?))?)*)[\x09\x0A\x0B\x0C\x0D\x20]*';
+    public const PCRE_HTML_ATTRIBUTE = '((?:[\x09\x0A\x0B\x0C\x0D\x20]+[^\x09\x0A\x0B\x0C\x0D\x20\x2F\x3E][^\x09\x0A\x0B\x0C\x0D\x20\x2F\x3D\x3E]*(?:[\x09\x0A\x0B\x0C\x0D\x20]*=[\x09\x0A\x0B\x0C\x0D\x20]*(?:"(?:[^"]*)"|\'(?:[^\']*)\'|(?:[^\x09\x0A\x0B\x0C\x0D\x20\x22\x27\x3E][^\x09\x0A\x0B\x0C\x0D\x20\x3E]*)?))?)*)[\x09\x0A\x0B\x0C\x0D\x20]*';
 
     /**
      * PCRE for XML attributes
      */
-    const PCRE_XML_ATTRIBUTE = '((?:\s+(?:(?:[^\s:]+:)?[^\s:]+)\s*=\s*(?:"(?:[^"]*)"|\'(?:[^\']*)\'))*)\s*';
+    public const PCRE_XML_ATTRIBUTE = '((?:\s+(?:(?:[^\s:]+:)?[^\s:]+)\s*=\s*(?:"(?:[^"]*)"|\'(?:[^\']*)\'))*)\s*';
 
     /**
      * XML Namespace
      */
-    const NAMESPACE_XML = 'http://www.w3.org/XML/1998/namespace';
+    public const NAMESPACE_XML = 'http://www.w3.org/XML/1998/namespace';
 
     /**
      * Atom 1.0 Namespace
      */
-    const NAMESPACE_ATOM_10 = 'http://www.w3.org/2005/Atom';
+    public const NAMESPACE_ATOM_10 = 'http://www.w3.org/2005/Atom';
 
     /**
      * Atom 0.3 Namespace
      */
-    const NAMESPACE_ATOM_03 = 'http://purl.org/atom/ns#';
+    public const NAMESPACE_ATOM_03 = 'http://purl.org/atom/ns#';
 
     /**
      * RDF Namespace
      */
-    const NAMESPACE_RDF = 'http://www.w3.org/1999/02/22-rdf-syntax-ns#';
+    public const NAMESPACE_RDF = 'http://www.w3.org/1999/02/22-rdf-syntax-ns#';
 
     /**
      * RSS 0.90 Namespace
      */
-    const NAMESPACE_RSS_090 = 'http://my.netscape.com/rdf/simple/0.9/';
+    public const NAMESPACE_RSS_090 = 'http://my.netscape.com/rdf/simple/0.9/';
 
     /**
      * RSS 1.0 Namespace
      */
-    const NAMESPACE_RSS_10 = 'http://purl.org/rss/1.0/';
+    public const NAMESPACE_RSS_10 = 'http://purl.org/rss/1.0/';
 
     /**
      * RSS 1.0 Content Module Namespace
      */
-    const NAMESPACE_RSS_10_MODULES_CONTENT = 'http://purl.org/rss/1.0/modules/content/';
+    public const NAMESPACE_RSS_10_MODULES_CONTENT = 'http://purl.org/rss/1.0/modules/content/';
 
     /**
      * RSS 2.0 Namespace
      * (Stupid, I know, but I'm certain it will confuse people less with support.)
      */
-    const NAMESPACE_RSS_20 = '';
+    public const NAMESPACE_RSS_20 = '';
 
     /**
      * DC 1.0 Namespace
      */
-    const NAMESPACE_DC_10 = 'http://purl.org/dc/elements/1.0/';
+    public const NAMESPACE_DC_10 = 'http://purl.org/dc/elements/1.0/';
 
     /**
      * DC 1.1 Namespace
      */
-    const NAMESPACE_DC_11 = 'http://purl.org/dc/elements/1.1/';
+    public const NAMESPACE_DC_11 = 'http://purl.org/dc/elements/1.1/';
 
     /**
      * W3C Basic Geo (WGS84 lat/long) Vocabulary Namespace
      */
-    const NAMESPACE_W3C_BASIC_GEO = 'http://www.w3.org/2003/01/geo/wgs84_pos#';
+    public const NAMESPACE_W3C_BASIC_GEO = 'http://www.w3.org/2003/01/geo/wgs84_pos#';
 
     /**
      * GeoRSS Namespace
      */
-    const NAMESPACE_GEORSS = 'http://www.georss.org/georss';
+    public const NAMESPACE_GEORSS = 'http://www.georss.org/georss';
 
     /**
      * Media RSS Namespace
      */
-    const NAMESPACE_MEDIARSS = 'http://search.yahoo.com/mrss/';
+    public const NAMESPACE_MEDIARSS = 'http://search.yahoo.com/mrss/';
 
     /**
      * Wrong Media RSS Namespace. Caused by a long-standing typo in the spec.
      */
-    const NAMESPACE_MEDIARSS_WRONG = 'http://search.yahoo.com/mrss';
+    public const NAMESPACE_MEDIARSS_WRONG = 'http://search.yahoo.com/mrss';
 
     /**
      * Wrong Media RSS Namespace #2. New namespace introduced in Media RSS 1.5.
      */
-    const NAMESPACE_MEDIARSS_WRONG2 = 'http://video.search.yahoo.com/mrss';
+    public const NAMESPACE_MEDIARSS_WRONG2 = 'http://video.search.yahoo.com/mrss';
 
     /**
      * Wrong Media RSS Namespace #3. A possible typo of the Media RSS 1.5 namespace.
      */
-    const NAMESPACE_MEDIARSS_WRONG3 = 'http://video.search.yahoo.com/mrss/';
+    public const NAMESPACE_MEDIARSS_WRONG3 = 'http://video.search.yahoo.com/mrss/';
 
     /**
      * Wrong Media RSS Namespace #4. New spec location after the RSS Advisory Board takes it over, but not a valid namespace.
      */
-    const NAMESPACE_MEDIARSS_WRONG4 = 'http://www.rssboard.org/media-rss';
+    public const NAMESPACE_MEDIARSS_WRONG4 = 'http://www.rssboard.org/media-rss';
 
     /**
      * Wrong Media RSS Namespace #5. A possible typo of the RSS Advisory Board URL.
      */
-    const NAMESPACE_MEDIARSS_WRONG5 = 'http://www.rssboard.org/media-rss/';
+    public const NAMESPACE_MEDIARSS_WRONG5 = 'http://www.rssboard.org/media-rss/';
 
     /**
      * iTunes RSS Namespace
      */
-    const NAMESPACE_ITUNES = 'http://www.itunes.com/dtds/podcast-1.0.dtd';
+    public const NAMESPACE_ITUNES = 'http://www.itunes.com/dtds/podcast-1.0.dtd';
 
     /**
      * XHTML Namespace
      */
-    const NAMESPACE_XHTML = 'http://www.w3.org/1999/xhtml';
+    public const NAMESPACE_XHTML = 'http://www.w3.org/1999/xhtml';
 
     /**
      * IANA Link Relations Registry
      */
-    const IANA_LINK_RELATIONS_REGISTRY = 'http://www.iana.org/assignments/relation/';
+    public const IANA_LINK_RELATIONS_REGISTRY = 'http://www.iana.org/assignments/relation/';
 
     /**
      * No file source
      */
-    const FILE_SOURCE_NONE = 0;
+    public const FILE_SOURCE_NONE = 0;
 
     /**
      * Remote file source
      */
-    const FILE_SOURCE_REMOTE = 1;
+    public const FILE_SOURCE_REMOTE = 1;
 
     /**
      * Local file source
      */
-    const FILE_SOURCE_LOCAL = 2;
+    public const FILE_SOURCE_LOCAL = 2;
 
     /**
      * fsockopen() file source
      */
-    const FILE_SOURCE_FSOCKOPEN = 4;
+    public const FILE_SOURCE_FSOCKOPEN = 4;
 
     /**
      * cURL file source
      */
-    const FILE_SOURCE_CURL = 8;
+    public const FILE_SOURCE_CURL = 8;
 
     /**
      * file_get_contents() file source
      */
-    const FILE_SOURCE_FILE_GET_CONTENTS = 16;
+    public const FILE_SOURCE_FILE_GET_CONTENTS = 16;
 
     /**
      * @var array Raw data
@@ -685,8 +688,8 @@ class SimplePie
      */
     public function __construct()
     {
-        if (version_compare(PHP_VERSION, '5.6', '<')) {
-            trigger_error('Please upgrade to PHP 5.6 or newer.');
+        if (version_compare(PHP_VERSION, '7.2', '<')) {
+            trigger_error('Please upgrade to PHP 7.2 or newer.');
             die();
         }
 
@@ -697,8 +700,7 @@ class SimplePie
         $this->registry = new \SimplePie\Registry();
 
         if (func_num_args() > 0) {
-            $level = defined('E_USER_DEPRECATED') ? E_USER_DEPRECATED : E_USER_WARNING;
-            trigger_error('Passing parameters to the constructor is no longer supported. Please use set_feed_url(), set_cache_location(), and set_cache_duration() directly.', $level);
+            trigger_error('Passing parameters to the constructor is no longer supported. Please use set_feed_url(), set_cache_location(), and set_cache_duration() directly.', \E_USER_DEPRECATED);
 
             $args = func_get_args();
             switch (count($args)) {
@@ -872,6 +874,18 @@ class SimplePie
     public function enable_cache($enable = true)
     {
         $this->cache = (bool) $enable;
+    }
+
+    /**
+     * Set a PSR-16 implementation as cache
+     *
+     * @param CacheInterface $psr16cache The PSR-16 cache implementation
+     */
+    public function set_cache(CacheInterface $cache)
+    {
+        Psr16::store_cache($cache);
+        $this->registry->call('Cache', 'register', ['psr16', Psr16::class]);
+        $this->cache_location = 'psr16';
     }
 
     /**
@@ -1416,7 +1430,7 @@ class SimplePie
             // Decide whether to enable caching
             if ($this->cache && $parsed_feed_url['scheme'] !== '') {
                 $filename = $this->get_cache_filename($this->feed_url);
-                $cache = $this->registry->call('Cache', 'get_handler', [$this->cache_location, $filename, 'spc']);
+                $cache = $this->registry->call('Cache', 'get_handler', [$this->cache_location, $filename, Base::TYPE_FEED]);
             }
 
             // Fetch the data via \SimplePie\File into $this->raw_data
@@ -1426,7 +1440,7 @@ class SimplePie
                 return false;
             }
 
-            list($headers, $sniffed) = $fetched;
+            [$headers, $sniffed] = $fetched;
         }
 
         // Empty response check
@@ -1759,7 +1773,7 @@ class SimplePie
                     if (!$cache->save($this)) {
                         trigger_error("$this->cache_location is not writable. Make sure you've set the correct relative or absolute path, and that the location is server-writable.", E_USER_WARNING);
                     }
-                    $cache = $this->registry->call('Cache', 'get_handler', [$this->cache_location, call_user_func($this->cache_name_function, $file->url), 'spc']);
+                    $cache = $this->registry->call('Cache', 'get_handler', [$this->cache_location, call_user_func($this->cache_name_function, $file->url), Base::TYPE_FEED]);
                 }
             }
             $this->feed_url = $file->url;
@@ -2924,8 +2938,7 @@ class SimplePie
      */
     public function set_favicon_handler($page = false, $qs = 'i')
     {
-        $level = defined('E_USER_DEPRECATED') ? E_USER_DEPRECATED : E_USER_WARNING;
-        trigger_error('Favicon handling has been removed, please use your own handling', $level);
+        trigger_error('Favicon handling has been removed, please use your own handling', \E_USER_DEPRECATED);
         return false;
     }
 
@@ -2936,8 +2949,7 @@ class SimplePie
      */
     public function get_favicon()
     {
-        $level = defined('E_USER_DEPRECATED') ? E_USER_DEPRECATED : E_USER_WARNING;
-        trigger_error('Favicon handling has been removed, please use your own handling', $level);
+        trigger_error('Favicon handling has been removed, please use your own handling', \E_USER_DEPRECATED);
 
         if (($url = $this->get_link()) !== null) {
             return 'https://www.google.com/s2/favicons?domain=' . urlencode($url);
@@ -2956,13 +2968,11 @@ class SimplePie
     public function __call($method, $args)
     {
         if (strpos($method, 'subscribe_') === 0) {
-            $level = defined('E_USER_DEPRECATED') ? E_USER_DEPRECATED : E_USER_WARNING;
-            trigger_error('subscribe_*() has been deprecated, implement the callback yourself', $level);
+            trigger_error('subscribe_*() has been deprecated, implement the callback yourself', \E_USER_DEPRECATED);
             return '';
         }
         if ($method === 'enable_xml_dump') {
-            $level = defined('E_USER_DEPRECATED') ? E_USER_DEPRECATED : E_USER_WARNING;
-            trigger_error('enable_xml_dump() has been deprecated, use get_raw_data() instead', $level);
+            trigger_error('enable_xml_dump() has been deprecated, use get_raw_data() instead', \E_USER_DEPRECATED);
             return false;
         }