+++ tests/CacheTest.php
@@ -41,6 +41,7 @@
  * @license http://www.opensource.org/licenses/bsd-license.php BSD License
  */
 
+use SimplePie\Tests\Fixtures\FileMock;
 use Yoast\PHPUnitPolyfills\Polyfills\ExpectPHPException;
 
 /**
@@ -80,7 +81,7 @@ class CacheTest extends PHPUnit\Framework\TestCase
 
     public function testDirectOverrideLegacy()
     {
-        if (PHP_VERSION_ID < 80000) {
+        if (version_compare(PHP_VERSION, '8.0', '<')) {
             $this->expectException('Exception_Success');
         } else {
             // PHP 8.0 will throw a `TypeError` for trying to call a non-static method statically.
@@ -91,7 +92,7 @@ class CacheTest extends PHPUnit\Framework\TestCase
 
         $feed = new SimplePie();
         $feed->set_cache_class('Mock_CacheLegacy');
-        $feed->get_registry()->register('File', 'MockSimplePie_File');
+        $feed->get_registry()->register('File', FileMock::class);
         $feed->set_feed_url('http://example.com/feed/');
 
         $feed->init();
@@ -103,7 +104,7 @@ class CacheTest extends PHPUnit\Framework\TestCase
 
         $feed = new SimplePie();
         $feed->get_registry()->register('Cache', 'Mock_CacheNew');
-        $feed->get_registry()->register('File', 'MockSimplePie_File');
+        $feed->get_registry()->register('File', FileMock::class);
         $feed->set_feed_url('http://example.com/feed/');
 
         $feed->init();